src/envs/promptops_arena/server/requirements.txt
openenv-core>=0.2.3
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
jsonschema>=4.20.0
transformers>=4.45.0
torch>=2.4.0
